More Cleanup (#299)
* uncomment some part of the unifier ??

* fix printing of meta

* simplify syntax code, make it more correct; optimize operations

Before, the open_var_as_cmd required that the 'cmd' argument be locally
closed, but it's not clear whether we were actually maintaining that
invariant in the code. (At least, it didn't seem to be violated, but...)

Now, I just use the 'subst' function to deal with this case, and get rid
of open_var_as_cmd. But in order for this to be performant, it was
necessary to store information about the maximum unbound de bruijn index
to appear in a term, at all the nodes. This optimization lets us stop
substituting on a branch as soon as we don't need to anymore.

Without the optimization, the naive version with subst was about 10x
slower in some hard examples (to do with univalence), but now it is just
as fast as it was before.

* abstract

* decrease use of evaluator etc in unifier